// examples/dataverse/src/generated/models/Siero_coursesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Siero_coursesstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Siero_coursesstatecode = keyof typeof Siero_coursesstatecode;
export const Siero_coursesstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Siero_coursesstatuscode = keyof typeof Siero_coursesstatuscode;

export interface Siero_courses {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  siero_courseid: string;
  siero_enddate?: string;
  siero_name: string;
  siero_startdate?: string;
  statecode: Siero_coursesstatecode;
  statecodename?: string;
  statuscode?: Siero_coursesstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  versionnumber?: string;
}