// examples/dataverse/src/generated/models/Siero_studentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Siero_studentsstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Siero_studentsstatecode = keyof typeof Siero_studentsstatecode;
export const Siero_studentsstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Siero_studentsstatuscode = keyof typeof Siero_studentsstatuscode;

export interface Siero_students {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  siero_courseid?: string;
  siero_courseidname?: string;
  siero_firstname?: string;
  siero_fullname?: string;
  siero_lastname?: string;
  siero_studentid: string;
  statecode: Siero_studentsstatecode;
  statecodename?: string;
  statuscode?: Siero_studentsstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  versionnumber?: string;
  _siero_courseid_value?: string;
}