// examples/dataverse/src/generated/models/Siero_subjectsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Siero_subjectsstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Siero_subjectsstatecode = keyof typeof Siero_subjectsstatecode;
export const Siero_subjectsstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Siero_subjectsstatuscode = keyof typeof Siero_subjectsstatuscode;

export interface Siero_subjects {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  siero_name: string;
  siero_subjectid: string;
  statecode: Siero_subjectsstatecode;
  statecodename?: string;
  statuscode?: Siero_subjectsstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  versionnumber?: string;
}